update Enemy
add more Enemy
